requirements.txt
streamlit==1.35.0
ultralytics>=8.3.0
opencv-python-headless==4.9.0.80
pillow==10.3.0
numpy==1.26.4
matplotlib==3.8.4
torch==2.3.1
torchvision==0.18.1
tensorflow-cpu==2.15.0
python-docx==1.2.0